--- design.txt
@@ -1,23 +1,0 @@
-Title : Flat mates Bill
-
-Description:
-
-An app that gets an input the amount if a bill for a particular period and the days that each of the flat
-mates stayed in the house for that period and returns how much each of them needs to pay.
-
-It also generates a pdf report stating the names of the flatmates, the period and how much each of them needs to pay.
-
-objects:
-
-Bill:
-    amount
-    period
-
-Flatmates:
-    name
-    days_in_house
-    pays(Bill)
-
-pdfReport:
-    filename
-    generate(flatmate1, flatmate2, bill)
