+++ files/design.txt
@@ -1,15 +1,23 @@
-Title: Flatmates Bill
-Description: An app that gets as input the amount of a bill for a particular period
-and the     days that each of the flatmates stayed in the house for that period
-and returns how much each flatmate has to pay. It also generates a PDF report
-stating the names of the flatmates, the period, and how much each of them had to pay.
-Objects: Bill:
-            amount
-            period
-         Flatmate:
-            name
-            days_in_house
-            pays(bill)
-         PdfReport:
-            filename
-            generate(flatmate1, flatmate2, bill)
+Title : Flat mates Bill
+
+Description:
+
+An app that gets an input the amount if a bill for a particular period and the days that each of the flat
+mates stayed in the house for that period and returns how much each of them needs to pay.
+
+It also generates a pdf report stating the names of the flatmates, the period and how much each of them needs to pay.
+
+objects:
+
+Bill:
+    amount
+    period
+
+Flatmates:
+    name
+    days_in_house
+    pays(Bill)
+
+pdfReport:
+    filename
+    generate(flatmate1, flatmate2, bill)
